organized machine bar file structure
