+++ src/App.svelte
@@ -1,11 +1,11 @@
 <script>
-	import ToolsBar from "./components/ToolsBar.svelte";
+	import MachineBar from "./components/MachinesBar/index.svelte";
 	import World from "./components/World.svelte";
 </script>
 
 <main>
-  <div class="flex flex-row h-screen">
-    <World />
-    <ToolsBar />
-  </div>
+	<div class="flex flex-row h-screen">
+		<World />
+		<MachineBar />
+	</div>
 </main>
