+++ src/components/MachinesBar/index.svelte
